requirements.txt
# EdgeChat Backend — we add packages as we go
# Step 0–1: app + config
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.0.0
# Step 2: PostgreSQL + SQLAlchemy async
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
# Step 3: Auth — password hashing + JWT (bcrypt directly; passlib has compat issues with newer bcrypt)
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
# Step 6: Redis for rate limiting
redis>=5.0.0
# Day 4: Gemini API
google-genai>=1.0.0
# Day 4 Step 2: MongoDB for chat history
motor>=3.3.0

# Day 5: Testing (+ optional .env.test for local test DB URLs)
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
python-dotenv>=1.0.0
